pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexo-crypto"
version = "0.0.1"
description = "Crypto package for Nexo"
authors = [
    { name = "Agra Bima Yuda", email = "agra@nexmedis.com" }
]
license = { text = "Proprietary"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types>=0.7.0",
    "bcrypt>=4.3.0",
    "black>=25.1.0",
    "cffi>=2.0.0",
    "cfgv>=3.4.0",
    "click>=8.2.1",
    "colorama>=0.4.6",
    "cryptography>=45.0.7",
    "distlib>=0.4.0",
    "filelock>=3.19.1",
    "identify>=2.6.14",
    "mypy_extensions>=1.1.0",
    "nexo-types>=0.0.1",
    "nodeenv>=1.9.1",
    "packaging>=25.0",
    "pathspec>=0.12.1",
    "platformdirs>=4.4.0",
    "pre_commit>=4.3.0",
    "pycparser>=2.23",
    "pycryptodome>=3.23.0",
    "pydantic>=2.11.9",
    "pydantic_core>=2.33.2",
    "PyJWT>=2.10.1",
    "PyYAML>=6.0.2",
    "typing-inspection>=0.4.1",
    "typing_extensions>=4.15.0",
    "virtualenv>=20.34.0",
]

[tool.setuptools]
packages = [
    "nexo.crypto",
    "nexo.crypto.encryption",
    "nexo.crypto.hash",
    "nexo.crypto.key",
    "nexo.crypto.key.rsa"
]

[tool.setuptools.package-data]
"nexo.crypto" = ["*.json", "*.yaml"]
"nexo.crypto.encryption" = ["*.json", "*.yaml"]
"nexo.crypto.hash" = ["*.json", "*.yaml"]
"nexo.crypto.key" = ["*.json", "*.yaml"]
"nexo.crypto.key.rsa" = ["*.json", "*.yaml"]

[tool.setuptools.package-dir]
"nexo.crypto" = "src"
"nexo.crypto.encryption" = "src/encryption"
"nexo.crypto.hash" = "src/hash"
"nexo.crypto.key" = "src/key"
"nexo.crypto.key.rsa" = "src/key/rsa"
